refactor: index.js 삭제
