+++ src/pages/Login.jsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 import styled from "styled-components";
-import Logo from "../../assets/logo.png";
-import LoginForm from "../../components/LoginForm";
+import Logo from "../assets/logo.png";
+import LoginForm from "../components/LoginForm";
 
 const Wrapper = styled.div`
   display: flex;
